EventConverter - moved to group with EventValues - As per pickle's request
