+++ src/main/java/ch/njol/skript/expressions/base/EventValueExpression.java
@@ -20,7 +20,7 @@ import ch.njol.skript.registrations.EventValues;
 import ch.njol.skript.util.Utils;
 import ch.njol.util.Kleenean;
 import ch.njol.util.coll.CollectionUtils;
-import org.skriptlang.skript.bukkit.event.EventConverter;
+import ch.njol.skript.registrations.EventConverter;
 import org.skriptlang.skript.registration.SyntaxInfo;
 import org.skriptlang.skript.registration.SyntaxRegistry;
 import org.skriptlang.skript.util.Priority;
